requirements.txt
Django==4.2.3
argon2-cffi==21.3.0
django-environ==0.10.0
Pillow==10.0.0
django-allauth==0.54.0
django-tailwind==3.6.0
django-crispy-forms==2.0
crispy-tailwind==0.5.0
django-countries
django-naomi
psycopg2==2.9.6
psycopg2-binary==2.9.6
django-widget-tweaks==1.4.12
django-browser-reload==1.11.0